Add a test framework. Small changes to diff behaviour.
Note that the test framework output can occassionally be polluted by output from the perforce extension
trying to "do it's own thing" -e.g. opening a window but so far this does not cause failing tests on my system.
Need to investigate whether something needs to be updated to prevent the extension interfering

The only functional change is:
 - INTEGRATE and BRANCH changes diff against the current file in perforce instead of the resolveFromFile

Other changes are minor bug fixes or useful non-functional changes to support testing.

The build target location has been changed from out/src to out/ as the src part was redundant / misleading

Squashed commits:

move assertions to a helper

[245f85a] add helper assertions for cleanliness and vastly improved errors - WIP

could be better as a plugin to chai

[64f8756] add test-fixtures to .vscodeignore

[18f4fe9] Make the Open function async. Add further tests for diffing

Open & Openvshelved now awaits the vscode.open and .diff commands

This allows us to wait for it to complete opening the files for diff (useful for unit testing)

[204f08c] change to returning tuple instead of providing callback.
working diff uri matcher.

[eff4170] Correct integration and branch left hand diff pane

[0e127e7] fstat updates (in progress)

[acd936b] fix test runner script

moves the test-fixtures out of src, to where it makes more sense

[b33f643] change from tdd to bdd style. add more tests

The bdd style provides better standard messages for stub calls

[a4b6741] update icon path to match build location change

[ebba4a1] custom behaviour for a changelist being shelved

[6556e06] tabs -> spaces

[c01f68c] Initial test framework & compile location

Provides a test framework for running tests using mocha, chai, + sinon for stubbing the perforce commands

Changed the build location to be out instead of out/src - which previously did not make much sense

Changed npm compile not to watch (otherwise we can't pretest by running compile)